method is not given in the keyword arguments,  so refer to self.method in BlogPost model save.   Add import of delete_post from blogs.views into blogs urls.py.     Import main models.users into blogs views.  Update the get_create_date method to account for the fact that the original date could be in history OR in the blog post itself.
Our USERNAME field in our custom User object is 'email'.  And, perhaps related,  when I try to assign request.user to blog_post.user (a Foriegn Key field under the covers),  it ends up trying to put an email into the field (which is a UUID field).   Which causes an error.  Trying to find ways around this.