If you add groups to the auth_group table,  you can now create new forum threads with groups,  and it will create new Threads/Forums with one selected ThreadRole associated with it
In forums/models.py:

Get rid of name field and  __str__ override method of ThreadRole.

In forums/views.py

Get rid of form_valid method

get() override now has the requesting user's roles  and an empty ThreadForm
in the context.

Add a post override.  Move some of the old form_valid logic here.  Extract the role from the POST and create a new ThreadRole with it.