ForumComment Rows will still be backed up even if 'ERROR' was assigned to the row_action.
List of Comments displayed under a forum post now display both the Creation Date/Original Author and the Last Edited Date/Editing Author of  each comment.

But if a comment has an incorrect value in its row_action cell,  a warning is displayed indicating instead the row might be a restored deleted row which displays the date of deletion rather than the date of the last edit.

And if a row with  row_action value = 'CREATE is not found in an *Edited* comments history,  a warning is also displayed